Update the profiler and add more util to protobuf library. #43

+++ lib-core/src/main/java/com/my/core/protocol/IProfiler.java
@@ -24,5 +24,5 @@ public interface IProfiler {
 
     void startProfiling();
 
-    long stopProfilingAndCalculateInterval();
+    float stopProfilingAndCalculateInterval();
 }
